requirements.txt
aiohttp==3.6.0
appdirs==1.4.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
chardet==3.0.4
Click==7.0
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
multidict==4.5.2
packaging==19.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
pytz==2019.2
six==1.12.0
toml==0.10.0
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0
